+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/MapboxNavigationNotification.java
@@ -10,27 +10,51 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.pm.PackageManager;
 import android.os.Build;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 import android.support.v4.content.ContextCompat;
 import android.text.SpannableString;
+import android.text.TextUtils;
 import android.text.format.DateFormat;
 import android.widget.RemoteViews;
 
 import com.mapbox.api.directions.v5.models.LegStep;
 import com.mapbox.api.directions.v5.models.RouteOptions;
+import com.mapbox.api.directions.v5.models.StepManeuver;
 import com.mapbox.navigator.BannerInstruction;
 import com.mapbox.services.android.navigation.R;
 import com.mapbox.services.android.navigation.v5.navigation.notification.NavigationNotification;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
 import com.mapbox.services.android.navigation.v5.utils.DistanceFormatter;
 import com.mapbox.services.android.navigation.v5.utils.LocaleUtils;
-import com.mapbox.services.android.navigation.v5.utils.ManeuverUtils;
 
 import java.util.Calendar;
 
 import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.NAVIGATION_NOTIFICATION_CHANNEL;
 import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.NAVIGATION_NOTIFICATION_ID;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_LEFT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_RIGHT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_SHARP_LEFT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_SHARP_RIGHT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_SLIGHT_LEFT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_STRAIGHT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_MODIFIER_UTURN;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_ARRIVE;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_CONTINUE;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_DEPART;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_END_OF_ROAD;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_FORK;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_MERGE;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_NEW_NAME;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_NOTIFICATION;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_OFF_RAMP;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_ON_RAMP;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_ROTARY;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_ROUNDABOUT;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_ROUNDABOUT_TURN;
+import static com.mapbox.services.android.navigation.v5.navigation.NavigationConstants.STEP_MANEUVER_TYPE_TURN;
 import static com.mapbox.services.android.navigation.v5.utils.time.TimeFormatter.formatTime;
 
 /**
@@ -123,6 +147,11 @@ class MapboxNavigationNotification implements NavigationNotification {
   // Package private (no modifier) for testing purposes
   String retrieveInstructionText() {
     return instructionText;
+  }
+
+  // Package private (no modifier) for testing purposes
+  int retrieveCurrentManeuverId() {
+    return currentManeuverId;
   }
 
   private void initialize(Context applicationContext, MapboxNavigation mapboxNavigation) {
@@ -271,17 +300,175 @@ class MapboxNavigationNotification implements NavigationNotification {
     expandedNotificationRemoteViews.setTextViewText(R.id.notificationArrivalText, time);
   }
 
-  private void updateManeuverImage(LegStep step) {
-    if (newManeuverId(step)) {
-      int maneuverResource = ManeuverUtils.getManeuverResource(step);
+  private void updateManeuverImage(@NonNull LegStep step) {
+    int maneuverResource = getManeuverResource(step);
+    if (currentManeuverId != maneuverResource) {
       currentManeuverId = maneuverResource;
       collapsedNotificationRemoteViews.setImageViewResource(R.id.maneuverImage, maneuverResource);
       expandedNotificationRemoteViews.setImageViewResource(R.id.maneuverImage, maneuverResource);
     }
   }
 
-  private boolean newManeuverId(LegStep step) {
-    return currentManeuverId != ManeuverUtils.getManeuverResource(step);
+  private int getManeuverResource(@NonNull LegStep step) {
+    StepManeuver maneuver = step.maneuver();
+    String maneuverType = maneuver.type();
+    String maneuverModifier = maneuver.modifier();
+    if (!TextUtils.isEmpty(maneuverModifier)) {
+      String drivingSide = step.drivingSide();
+      if (isLeftDrivingSideAndRoundaboutOrRotaryOrUturn(maneuverType, maneuverModifier, drivingSide)) {
+        return obtainManeuverResourceFrom(maneuverType + maneuverModifier + drivingSide);
+      }
+      return obtainManeuverResourceFrom(maneuverType + maneuverModifier);
+    }
+    return obtainManeuverResourceFrom(maneuverType);
+  }
+
+  private int obtainManeuverResourceFrom(String maneuver) {
+    switch (maneuver) {
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_UTURN:
+      case STEP_MANEUVER_TYPE_CONTINUE + STEP_MANEUVER_MODIFIER_UTURN:
+        return R.drawable.ic_maneuver_turn_180;
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_UTURN + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_CONTINUE + STEP_MANEUVER_MODIFIER_UTURN + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_turn_180_left_driving_side;
+
+      case STEP_MANEUVER_TYPE_ARRIVE + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_arrive_left;
+      case STEP_MANEUVER_TYPE_ARRIVE + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_arrive_right;
+      case STEP_MANEUVER_TYPE_ARRIVE:
+        return R.drawable.ic_maneuver_arrive;
+
+      case STEP_MANEUVER_TYPE_DEPART + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_depart_left;
+      case STEP_MANEUVER_TYPE_DEPART + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_depart_right;
+      case STEP_MANEUVER_TYPE_DEPART:
+        return R.drawable.ic_maneuver_depart;
+
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_SHARP_RIGHT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_SHARP_RIGHT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_SHARP_RIGHT:
+        return R.drawable.ic_maneuver_turn_75;
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_RIGHT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_RIGHT:
+      case STEP_MANEUVER_TYPE_ROUNDABOUT_TURN + STEP_MANEUVER_MODIFIER_RIGHT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_turn_45;
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+        return R.drawable.ic_maneuver_turn_30;
+
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_SHARP_LEFT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_SHARP_LEFT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_SHARP_LEFT:
+        return R.drawable.ic_maneuver_turn_75_left;
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROUNDABOUT_TURN + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_turn_45_left;
+      case STEP_MANEUVER_TYPE_TURN + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+      case STEP_MANEUVER_TYPE_ON_RAMP + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+        return R.drawable.ic_maneuver_turn_30_left;
+
+      case STEP_MANEUVER_TYPE_MERGE + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_MERGE + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+        return R.drawable.ic_maneuver_merge_left;
+      case STEP_MANEUVER_TYPE_MERGE + STEP_MANEUVER_MODIFIER_RIGHT:
+      case STEP_MANEUVER_TYPE_MERGE + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+        return R.drawable.ic_maneuver_merge_right;
+
+      case STEP_MANEUVER_TYPE_OFF_RAMP + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_off_ramp_left;
+      case STEP_MANEUVER_TYPE_OFF_RAMP + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+        return R.drawable.ic_maneuver_off_ramp_slight_left;
+
+      case STEP_MANEUVER_TYPE_OFF_RAMP + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_off_ramp_right;
+      case STEP_MANEUVER_TYPE_OFF_RAMP + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+        return R.drawable.ic_maneuver_off_ramp_slight_right;
+
+      case STEP_MANEUVER_TYPE_FORK + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_fork_left;
+      case STEP_MANEUVER_TYPE_FORK + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+        return R.drawable.ic_maneuver_fork_slight_left;
+      case STEP_MANEUVER_TYPE_FORK + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_fork_right;
+      case STEP_MANEUVER_TYPE_FORK + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+        return R.drawable.ic_maneuver_fork_slight_right;
+      case STEP_MANEUVER_TYPE_FORK + STEP_MANEUVER_MODIFIER_STRAIGHT:
+        return R.drawable.ic_maneuver_fork_straight;
+      case STEP_MANEUVER_TYPE_FORK:
+        return R.drawable.ic_maneuver_fork;
+
+      case STEP_MANEUVER_TYPE_END_OF_ROAD + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_end_of_road_left;
+      case STEP_MANEUVER_TYPE_END_OF_ROAD + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_end_of_road_right;
+
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_left;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SHARP_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SHARP_LEFT:
+        return R.drawable.ic_maneuver_roundabout_sharp_left;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT:
+        return R.drawable.ic_maneuver_roundabout_slight_left;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_RIGHT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_RIGHT:
+        return R.drawable.ic_maneuver_roundabout_right;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SHARP_RIGHT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SHARP_RIGHT:
+        return R.drawable.ic_maneuver_roundabout_sharp_right;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT:
+        return R.drawable.ic_maneuver_roundabout_slight_right;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_STRAIGHT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_STRAIGHT:
+        return R.drawable.ic_maneuver_roundabout_straight;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT:
+      case STEP_MANEUVER_TYPE_ROTARY:
+        return R.drawable.ic_maneuver_roundabout;
+
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_left_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SHARP_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SHARP_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_sharp_left_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SLIGHT_LEFT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_slight_left_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_right_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SHARP_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SHARP_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_sharp_right_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_SLIGHT_RIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_slight_right_left_driving_side;
+      case STEP_MANEUVER_TYPE_ROUNDABOUT + STEP_MANEUVER_MODIFIER_STRAIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+      case STEP_MANEUVER_TYPE_ROTARY + STEP_MANEUVER_MODIFIER_STRAIGHT + STEP_MANEUVER_MODIFIER_LEFT:
+        return R.drawable.ic_maneuver_roundabout_straight_left_driving_side;
+
+      case STEP_MANEUVER_TYPE_MERGE + STEP_MANEUVER_MODIFIER_STRAIGHT:
+      case STEP_MANEUVER_TYPE_NOTIFICATION + STEP_MANEUVER_MODIFIER_STRAIGHT:
+      case STEP_MANEUVER_TYPE_CONTINUE + STEP_MANEUVER_MODIFIER_STRAIGHT:
+      case STEP_MANEUVER_TYPE_NEW_NAME + STEP_MANEUVER_MODIFIER_STRAIGHT:
+      default:
+        return R.drawable.ic_maneuver_turn_0;
+    }
+  }
+
+  private boolean isLeftDrivingSideAndRoundaboutOrRotaryOrUturn(String maneuverType, String maneuverModifier,
+                                                                String drivingSide) {
+    return STEP_MANEUVER_MODIFIER_LEFT.equals(drivingSide) && (STEP_MANEUVER_TYPE_ROUNDABOUT.equals(maneuverType)
+      || STEP_MANEUVER_TYPE_ROTARY.equals(maneuverType) || STEP_MANEUVER_MODIFIER_UTURN.equals(maneuverModifier));
   }
 
   private void onEndNavigationBtnClick() {